refactor: extract contributors mdx to child component
